requirements.txt
soprano-tts
torch
